+++ src/mathlive.ts
@@ -133,7 +133,7 @@ export function renderMathInElement(
   _renderMathInElement(el, optionsPrivate);
 }
 
-export { initVirtualKeyboardInCurrentBrowsingContext } from "virtual-keyboard/global";
+export { mountMathVirtualKeyboard as initVirtualKeyboardInCurrentBrowsingContext } from 'virtual-keyboard/global';
 
 /**
  * Current version: `{{SDK_VERSION}}`